添加会员重置密码接口

+++ src/members/members.service.ts
@@ -4,12 +4,12 @@
  * @email: 969718197@qq.com
  * @github: https://github.com/z-xuanyu
  * @Date: 2021-08-10 16:27:31
- * @LastEditTime: 2021-08-12 14:16:24
+ * @LastEditTime: 2021-09-03 11:44:31
  * @Description: Modify here please
  */
 import { Member } from '@libs/db/models/member.model';
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
-import { ReturnModelType } from '@typegoose/typegoose';
+import { mongoose, ReturnModelType } from '@typegoose/typegoose';
 import { InjectModel } from 'nestjs-typegoose';
 import { TableResponseResult } from 'src/BaseResponseResult';
 import { AddMemberDto } from './dto/addMember.dto';
@@ -19,11 +19,11 @@ import { UpdateMemberDto } from './dto/updateMember.dto';
 @Injectable()
 export class MembersService {
     constructor(@InjectModel(Member)
-    private readonly memberModel: ReturnModelType<typeof Member>){}
+    private readonly memberModel: ReturnModelType<typeof Member>) { }
 
 
     // 获取会员列表
-    async getMembers(parameters:GetMembersDto):Promise<TableResponseResult<Member>>{
+    async getMembers(parameters: GetMembersDto): Promise<TableResponseResult<Member>> {
         const total = await this.memberModel.countDocuments();
         const result = await this.memberModel.find({ name: { $regex: new RegExp(parameters.name, 'i') } }).limit(~~parameters.pageSize).skip(~~((parameters.pageNumber - 1) * parameters.pageSize))
         return {
@@ -33,7 +33,7 @@ export class MembersService {
     }
 
     // 添加会员
-    async addMember(parameters:AddMemberDto):Promise<Member>{
+    async addMember(parameters: AddMemberDto): Promise<Member> {
         const result = await this.memberModel.create(parameters as Member)
         if (!result) {
             throw new HttpException('系统异常，请联系管理员', HttpStatus.OK);
@@ -43,7 +43,12 @@ export class MembersService {
 
 
     // 编辑会员
-    async updateMember(id: string,parameters:UpdateMemberDto):Promise<Member>{
+    async updateMember(id: string, parameters: UpdateMemberDto): Promise<Member> {
+        // 认证会员id是否正确
+        const isObjectId = mongoose.Types.ObjectId.isValid(id)
+        if (!isObjectId) {
+            throw new HttpException('会员Id有误!', HttpStatus.OK);
+        }
         const result = await this.memberModel.findByIdAndUpdate(id, parameters)
         if (!result) {
             throw new HttpException('系统异常，请联系管理员', HttpStatus.OK);
@@ -53,8 +58,27 @@ export class MembersService {
 
 
     // 删除会员
-    async delMember(id:string):Promise<Member>{
+    async delMember(id: string): Promise<Member> {
+        // 认证会员id是否正确
+        const isObjectId = mongoose.Types.ObjectId.isValid(id)
+        if (!isObjectId) {
+            throw new HttpException('会员Id有误!', HttpStatus.OK);
+        }
         const result = await this.memberModel.findByIdAndDelete(id);
+        if (!result) {
+            throw new HttpException('系统异常，请联系管理员', HttpStatus.OK);
+        }
+        return result;
+    }
+
+    // 重置会员密码
+    async resetMemberPassword(id: string, newPassword: string): Promise<Member> {
+        // 认证会员id是否正确
+        const isObjectId = mongoose.Types.ObjectId.isValid(id)
+        if (!isObjectId) {
+            throw new HttpException('会员Id有误!', HttpStatus.OK);
+        }
+        const result = await this.memberModel.findByIdAndUpdate(id, { password: newPassword })
         if (!result) {
             throw new HttpException('系统异常，请联系管理员', HttpStatus.OK);
         }
